# js_project/server/config/swagger.yaml
openapi: 3.0.0
# Added by API Auto Mocking Plugin
servers:
  - description: SwaggerHub API Auto Mocking
    url: ''
info:
  description: This is a simple API
  version: "1.0.0"
  title: Simple Inventory API
  contact:
    email: you@your-company.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
tags:
  - name: Articles
    description: Some list
paths:
  /api/list:
    get:
      tags:
        - Articles
      summary: searches serverData
      operationId: searchserverData
      description: search serverData list
      responses:
        '200':
          description: search results matching criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  required:
                    - id
                    - text
                  properties:
                    header:
                      type: object
                      properties:
                        headTitle:
                          type: string
                          example: Ideas and perspectives you won’t find...
                        headText: 
                          type: string
                          example: BlogWorld taps into the brains of the world’s...
                    about:
                      type: object
                      properties:
                        headingTitle:
                          type: string
                          example: About us
                        headingText: 
                          type: string
                          example: This is who we are
        '400':
          description: bad input parameter
  /api/articles:
    get:
      tags:
        - Articles
      summary: searches all articles
      operationId: searchArticle
      description: search articles list
      responses:
        '200':
          description: search results matching criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  required:
                    - id
                    - text
                  properties:
                    id:
                      type: string
                      format: uuid
                      example: d290f1ee-6c54-4b01-90e6-d701748f0851
                    name:
                      type: string
                      example: Neil Richards
                    postType:
                      type: string
                      example: Video
                    title:
                      type: string
                      example: In the Future We Will All Live in Star Wars
                    text:
                      type: string
                      example: New text for article
        '400':
          description: bad input parameter
  /api/articles/{id}:
    get:
      tags:
        - Articles
      summary: searches article by id
      operationId: searchArticleByID
      parameters:
       - name: id
         in: path
         description: ID of article to return
         required: true
         schema:
            type: integer
            format: int64
      description: search article by ID
      responses:
        '200':
          description: search results matching criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  required:
                    - id
                    - text
                  properties:
                    id:
                      type: string
                      format: uuid
                      example: d290f1ee-6c54-4b01-90e6-d701748f0851
                    name:
                      type: string
                      example: Neil Richards
                    postType:
                      type: string
                      example: Video
                    title:
                      type: string
                      example: In the Future We Will All Live in Star Wars
                    text:
                      type: string
                      example: New text for article
        '400':
          description: bad input parameter
  /api/deleteArticles:
    delete:
      tags:
        - Articles
      summary: delete all articles
      operationId: deleteArticle
      description: delete article lists
      responses:
        '200':
          description: all articles have been deleted
        '400':
          description: bad input parameter
  /api/articles/{id}:
    delete:
      tags:
        - Articles
      summary: delete article by id
      operationId: deleteArticleByID
      parameters:
       - name: id
         in: path
         description: ID of article to return
         required: true
         schema:
            type: integer
            format: int64
      description: delete article by ID
      responses:
        '200':
          description: search results matching criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Article'
        '400':
          description: bad input parameter
  /api/create-article:
    post:
      tags:
        - Articles
      summary: add new article
      operationId: addArticle
      description: Add new article
      responses:
        '201':
          description: item created
        '400':
          description: 'invalid input, object invalid'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - id
                - text
              properties:
                id:
                  type: string
                  format: uuid
                  example: d290f1ee-6c54-4b01-90e6-d701748f0851
                name:
                  type: string
                  example: Neil Richards
                postType:
                  type: string
                  example: video
                title:
                  type: string
                  example: In the Future We Will All Live in Star Wars
                text:
                  type: string
                  example: New text for article
        description: Article to add
  /api/create-article/{id}:
    put:
      tags:
        - Articles
      summary: update article
      operationId: updateArticleByID
      parameters:
       - name: id
         in: path
         description: ID of article to return
         required: true
         schema:
            type: integer
            format: int64
      description: update article by ID
      responses:
        '201':
          description: the article has been updated
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  required:
                    - id
                    - text
                  properties:
                    id:
                      type: string
                      format: uuid
                      example: d290f1ee-6c54-4b01-90e6-d701748f0851
                    text:
                      type: string
                      example: New text for article
                    name:
                      type: string
                      example: Neil Richards
                    postType:
                      type: string
                      example: Video
                    title:
                      type: string
                      example: In the Future We Will All Live in Star Wars
        '400':
          description: 'invalid input, object invalid'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - text
              properties:
                text:
                  type: string
                  example: New text for article
        description: Article to update
components:
  schemas:
    Article:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
          format: uuid
          example: d290f1ee-6c54-4b01-90e6-d701748f0851
        name:
          type: string
          example: Widget Adapter